refactor[mobile]: move screens to root dir
