+++ mobile/components/index.ts
@@ -1,1 +1,13 @@
+export * from "./AppBottomSheet";
+export * from "./AuthWrapper";
+export * from "./BottomTabBar";
+export * from "./Collapsible";
+export * from "./ExternalLink";
+export * from "./Gutter";
+export * from "./Header";
+export * from "./IconRound";
+export * from "./ParallaxScrollView";
+export * from "./Row";
 export * from "./Wrapper";
+export * from "./TText";
+export * from "./TView";
