+++ mobile/screens/market/Assets/AssetsCrypto.tsx
