+++ mobile/screens/market/Assets/AssetsRow.tsx
